requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
streamlit==1.28.1
requests==2.31.0
plotly==5.17.0
streamlit-autorefresh==1.0.0
python-dotenv==1.0.0
